add missing copy assignment
Add missing copy assignment operator if a copy constructor is defined.
Made the copy constructor, destructor, and copy assignment be the
default compiler generated version if possible (or deleted them to use
the implicitly generated versions if they were not exported).

The inferiorMap compiler generated version was not correct as it created
copies of pointers that left dangling pointers when the inferiorMap  was
deleted.