Fix crash processing separate debug ELF files
When processing some separate debug ELF files, Dyninst can crash
because it creates and dereferences pointers based on the contents of
the ELF section headers.  In a separate debug file most non-debug
section's is removed from the ELF file, but the section header is left
unchanged with the exception that its type is changed to NOBITS.
Fixed by changing the processing of ELF files to treat NOBIT sections
as if they do not exist, eliminating Dyninst's reading via the invalid
memory location.